Move a few things to public API
