+++ dbos/__init__.py
@@ -1,3 +1,17 @@
 from .context import SetWorkflowUUID as SetWorkflowUUID
 from .dbos import DBOS as DBOS
+from .dbos import DBOSConfiguredInstance as DBOSConfiguredInstance
 from .dbos_config import ConfigFile as ConfigFile
+from .system_database import WorkflowStatusString as WorkflowStatusString
+from .workflow import WorkflowHandle as WorkflowHandle
+from .workflow import WorkflowStatus as WorkflowStatus
+
+__all__ = [
+    "ConfigFile",
+    "DBOS",
+    "DBOSConfiguredInstance",
+    "SetWorkflowUUID",
+    "WorkflowHandle",
+    "WorkflowStatus",
+    "WorkflowString",
+]
