+++ tests/scheduler/test_scheduler.py
@@ -1,6 +1,7 @@
 import time
 from datetime import datetime
 
+# Public API
 from dbos import DBOS
