+++ tests/test_concurrency.py
@@ -6,6 +6,7 @@ from typing import Tuple
 
 from sqlalchemy import text
 
+# Public API
 from dbos import DBOS, SetWorkflowUUID
