+++ tests/test_package.py
@@ -9,6 +9,7 @@ import urllib.request
 
 import sqlalchemy as sa
 
+# Private API because this is a test
 from dbos.dbos_config import load_config
