+++ tests/test_schema_migration.py
@@ -5,7 +5,10 @@ import sqlalchemy as sa
 from alembic import command
 from alembic.config import Config
 
+# Public API
 from dbos import DBOS, ConfigFile
+
+# Private API because this is a unit test
 from dbos.schemas.system_database import SystemSchema
